First commit of vba module
This module contains subroutines for: 1) creating and organizing csv input files into folders on a user's local drive or shared drive for each client entity, 2) loops through client entity folders to copy csv input files to an Excel report template containing dashboards, and 3) finally saving only the dashboard sheets to a final pdf for distribution to the client. This module was built to generate many standard pdf reports for external distribution to clients.